Add files via upload
Basic Queue Management System I made using PHP, the code is messy and I have not worked on it in a while, but the basic functionality must remain. Used XAMPP to code this. To use this put the folder in XAMPP's htdocs directory and use the apache dashboard to go to localhost/QMS/index.php